fix norm

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: jihong <jihong@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/13 16:31:57 by jihong            #+#    #+#             */
-/*   Updated: 2022/01/11 22:17:01 by jihong           ###   ########.fr       */
+/*   Updated: 2022/01/12 18:32:17 by jihong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 
 char	*get_next_line(int fd);
 size_t	ft_strlen(const char *c);
-char	*ft_strndup(char *str, int n);
+char	*ft_strndup(char *str, size_t n);
 int	check_newline(char *str);
 char	*ft_strjoin(char *str1, char *str2);
 char	*ft_strrchr(const char *str, int c);
